requirements.txt
astroid==1.6.1
dj-database-url==0.5.0
Django==2.0.2
django-heroku==0.3.1
djangorestframework==3.7.7
flake8==3.5.0
gunicorn==19.7.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.2
python-decouple==3.1
pytz==2018.3
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
